requirements.txt
Pillow==10.3.0
PlexAPI==4.15.12
Requests==2.31.0
python-dotenv==1.0.1
praw==7.8.1